server/requirements.txt
flask
flask-cors
flask-jwt-extended
flask-sqlalchemy
Flask-Migrate
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
google-genai
sqlalchemy
alembic
psycopg2-binary
google-genai
python-dotenv
mcp
PyPDF2
python-docx
openpyxl
werkzeug
scikit-learn
